Split actions into separate file
